fix #3901
